atari/requirements.txt
absl-py==0.8.0
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.0
beautifulsoup4==4.6.0
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
cloudpickle==1.1.1
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
Cython==0.26.1
decorator==4.4.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
dominate==2.4.0
easydict==1.9
gast==0.2.2
gin-config==0.1.3
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
helpdev==0.6.10
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
importlib-metadata==0.20
ipython==5.5.0
ipython-genutils==0.2.0
kazam==1.4.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
leveldb==0.1
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Markdown==3.1.1
matplotlib==2.1.1
more-itertools==7.2.0
netifaces==0.10.4
networkx==1.11
nose==1.3.7
numexpr==2.6.4
numpy==1.17.1
oauth==1.0.1
olefile==0.45.1
opencv-python==4.1.1.26
opt-einsum==3.0.1
pandas==0.22.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
protobuf==3.9.1
psutil==5.6.3
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.2.0
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
python-apt==1.6.4
python-dateutil==2.8.0
python-debian==0.1.32
python-gflags==3.1.2
pytz==2018.3
PyWavelets==0.5.1
pyxdg==0.25
PyYAML==3.12
QDarkStyle==2.7
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scikit-image==0.13.1
scipy==0.19.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplegeneric==0.8.1
simplejson==3.13.2
six==1.12.0
system-service==0.3
systemd-python==234
tables==3.4.2
tb-nightly==1.15.0a20190903
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==2.0.0rc0
termcolor==1.1.0
tf-agents-nightly==0.2.0.dev20190904
tf-estimator-nightly==1.14.0.dev2019090401
tf-nightly-gpu==1.15.0.dev20190730
tfp-nightly==0.9.0.dev20190904
Theano==1.0.4+19.g3be13b255
tqdm==4.35.0
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5
Werkzeug==0.15.6
wrapt==1.11.2
xkit==0.0.0
zipp==0.6.0
zope.interface==4.3.2
